Changed a bit how things work.
Rather than reusing UIViewControllers (TerminalViewControllers), we
just flush those down the toilet and treat them as renewable
resources.

Instead, what I do is I preserve the SshTerminalView, which actually
contains the terminal and the conenction state, and then I rehost
those on fresh view controllers on demand.

This eliminated some ugly bugs I had dealing with the view frames that
I could not get right, and perhaps assorted side effects.

Also cleaned up the observer for keyboard events while I was at it.

This change alone brought a lot of stability and I can actually use this
on the phone edition.   Still a few kinks to work out with the keyboard,
but I like where this is going.