Host is now a class, and not a structure.  This will allow me to make changes to it, and then post a notification, without worrying about copies being out of sync.
Additionally, there is a logical breakup between features of Host that
can be changed at runtime from things that can not be changed.  For
the latter, client code can post a notification, which can be tracked,
and updated.

The AppTerminalView no longer takes ad-hoc parameters, and instead takes
a Host, which allows us to track the changes that might be done by a
configurable terminal.

This reduces some ugly hacks in constructors for both SshTerminalView
and AppTerminalView.

MetalHost now tracks the fragment name that was used, should be useful
to prevent gratuitous recreation of content.

BackgroundSelector should be tracking the name of the background,
which it was not (previously it was necessary to track both the
background and the code).

ConfigurableUITerminal is now the entry point for terminals (as
opposed to the rawer SwiftUITerminal), as this one will provide both
the setting icon at the top, and provide the configuration dialog, and
should really be used across.

Still lots of bugs on the actual application of the terminal, and
currently, it is only used in one place, we need to put it in more
places (not on the Sessions view though).