This took me more time than I hoped.
The problem with the switching from Solid colors to Metal, and among
Metal shaders was caused by a few subtle issues.

First, the startup sequence for the TerminalViewControler would insert
the CAMetalLayer containing our target into the
ViewController.view.layer at position 0, but only during viewDidLoad().

This means that other attempts later, if I created a new layer, and
injected it, would either not get injected.

So now, when creating a layer when updating the live background, there
are two conditions: either the AppTerminalView is not currently
attached, so we just created the CAMetalLayer (and wait for something
else to position it) or we are attached, in which case, we need to
both set the frame for the layer, as well as attaching it to the
superview's layer.

This solves most switching between Live and Metal issues.