HostEditView: With the move of Host from struct to class, this broke the Host editor, so use a more explicit and cleaner path: rather than rely on the binding on the Host, we copy the data initially, and copy the data back into the Host - this is a good practice for editors anyways (we were doing this for a handful of things anyways, and was inconsistent).
Change the ellipsis for a pencil to edit the host configuration, I am
not a fan of it, but I think it is more clear what this does.

SessionsView: replace the SwiftUITerminal with a
ConfiguredReusedTerminal which gives us all the config goods.

ConfigurableTerminal: Rather than doing all the legwork to have two
constructors (that was looking uglier and uglier), replicate the code,
it is not that bad.

For now, since I have not done the proper workflow for ensuring a
username, I do not let a Host be saved without a username