SecureEnclave keys work.
Generally, they worked, but I kept reusing the same name, and this did
not delete the old value, so the public key generated corresponded to
a different key than the one that had been created.

Now, I store in the keychain using a name with a UUID.

The key display also no longer displays the private key or password
for the secure enclave keys (or the ECDSA key, which we currently do
not password protect).

Additionally, when removing keys from the UI, we remove the backing
key from the keychain.

That said, who is this "we" I keep talking about?  It was me, I did
the change, yet it feels strange to write a commit entry where I state
that I did remove the key from the keychain, it almost feels like
there was some teamwork, between the compiler, the simulator, the
device and me.  Good work team.