Channel: instead of a two-stage configuration for the readCallback, set this up in the constructor.
Channel: rather than returning true/false status conditions, return
the libssh2 status code, so we can provide better information in the
future.

Channel: send assumes that it is being called from the UI queue,
and does no longer use a separate queue to call into libssh2 to
call channel_write

SshTerminalView: split the channel reading operation into its own
method for clarity.  Also allows for data=nil && errorData=nil in
situations where we are invoked purely by the channel.ping (this is
different than the original SwiftSH sample code, and this is because
the libssh2 currently has a design limitation in that data it is not
possible to tell which channel got data when data is received on the
session, so we need to ping all available channels.