The code was calling libssh2_channel_eof from the wrong thread, need to ensure that this only happens from the actor.
Introduced a receivedEofUnsafe versiont that can be invoked from the
Channel read callback.

I do not know if two awaits to an actor incur a price cost, but just
to be safe, I mixed both the ping() with the eof status return, so
that pings on the channel also return whether the channel is done, so
we do not await twice for a single trip to the actor.

Should fix a crash I had been chasing down for a while and was just
reported.