Do not queue another network operation if we receive an error.
Previously the error was only bubbled up to return to the other
end of the network processing (our recv_ method callback).   While
this is great, and it worked in terms of the contract with libssh2,
this still queued another IO operation by calling itself recursively.

What this now does is stop the queueing of another network operation
receive operation if the operation has faulted.

Before this, if you closed a bunch of sessions, it would eventually
degrade the performance of the app until it was too slow.