More work on CloudKit/CoreData syncing
Many of the changes are related to CloudKit and CoreData, but some are
related to how I am storing the information on the Keychain.

1. HostEditView rather than having a plain "var host", which was wrong
to begin with, now uses a "@State var host: Host?" for storage.  It
really should be a @Binding, so that the value could be reflected on
the caller, but I could not find a way of having a Binding that could
have a nullable value.

This is necessary because the "Add" operation would before rely on an
empty "Host" object, and then edit this.  The problem is that in the
old world, if the user canceled, the Host was thrown away.  But in the
new CoreData world, every Host created is an object added to the
database, so we get visual glitches like the host is shown, before the
display pops, and if the user cancels, we would need to delete it -
but also, the Host would exist in a broken state if synced behind our
back.

2. Solving the pops up/pops down and also the repaint on the container
were solved by issuing first the call to dismiss(), then, the call to
host.objectWillChange.send(), this prevents the scenario where we
changed, which would trigger a refresh of our own view (since we have
a .sheet presented based on the host being edited).

3. Implemented Delete for hosts properly.

4. Added the sId UUID to the database for the Host records, not sure
why I had not done it.

KeyChain on iCloud:

Passowrds stored on the keychain and referenced from the Hosts will
now be synced via iCloud.

KeyChain on iCloud will not sync from simulator to devices, it only
syncs from device to device, so that took a while to figure out.

Keychain considers the kSecAttrSynchronizable on the query as part of
the key for looking up, so if this is not set, the value is not shown
at all, so I was saving with it, but failing to retrieve it.

Lastly, Keychain on iCloud does not sync immediately, hard to tell when it does.