Rename the CoreData Host to CHost, to follow the existing pattern everywhere else, but this allows me to introduce a Host protocol, as well as a companion MemoryHost class.
Both CHost and MemoryHost conform to the Host protocol.

The SwiftUI code that deals with editing and showing the collection of
hosts, uses the CoreData information, while the code that must connect
relies purely on the protocol.

This enables two features that were removed during the migration, the
URL open link, and the quick launch.

Still pending: recent hosts display, before I am done.