Move the keys to CoreData/CloudKit as well
Create a companion Key protocol, and then we have two implementations
of it CKey (the CloudData version) and MemoryKey (for transient
operations).

Split the saveKeychainElements from the original Key into direct APIs
that talk directly to the keychain

Implemented support for using the database to lookup URLs to connect
to, rather than the old Json system.

Keytools.reset: convenience function to clear all the keychain elements.

Keytools.dump: convenice function to dump all the contents from the keychain.

Completes some loose ends on the Host migration as well.